# draco-nodejs/backend/openapi.yaml
openapi: 3.0.0
info:
  title: Draco API
  version: 1.0.0
servers:
  - url: https://localhost:3001
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TeamManager:
      type: object
      properties:
        id:
          type: string
          pattern: ^d+$
        teamSeasonId:
          type: string
          pattern: ^d+$
        contact:
          type: object
          properties:
            id:
              type: string
              pattern: ^d+$
            firstName:
              type: string
              minLength: 1
              maxLength: 50
            lastName:
              type: string
              minLength: 1
              maxLength: 50
            middleName:
              type: string
              maxLength: 50
            email:
              type: string
              maxLength: 100
              format: email
            userId:
              type: string
              maxLength: 50
            photoUrl:
              type: string
              format: uri
            contactDetails:
              type: object
              properties:
                phone1:
                  type: string
                  nullable: true
                  default: ""
                phone2:
                  type: string
                  nullable: true
                  default: ""
                phone3:
                  type: string
                  nullable: true
                  default: ""
                streetaddress:
                  type: string
                  nullable: true
                  maxLength: 100
                  default: ""
                city:
                  type: string
                  nullable: true
                  maxLength: 50
                  default: ""
                state:
                  type: string
                  nullable: true
                  maxLength: 50
                  default: ""
                zip:
                  type: string
                  nullable: true
                  maxLength: 10
                  default: ""
                dateofbirth:
                  type: string
                  nullable: true
                  default: ""
          required:
            - id
            - firstName
            - lastName
      required:
        - id
        - teamSeasonId
        - contact
      description: Schema for a manager of a team for a given season
    CreateTeamManager:
      type: object
      properties:
        contact:
          type: object
          properties:
            id:
              type: string
              pattern: ^d+$
          required:
            - id
      required:
        - contact
      description: Schema for creating a manager
...
    ValidationError:
      type: object
      properties:
        message:
          type: string
        statusCode:
          type: number
          enum:
            - 400
        isRetryable:
          type: boolean
          enum:
            - false
      required:
        - message
        - statusCode
        - isRetryable
      description: Validation error (400)
      example:
        message: Validation failed
        statusCode: 400
        isRetryable: false
    AuthenticationError:
      type: object
      properties:
        message:
          type: string
        statusCode:
          type: number
          enum:
            - 401
        isRetryable:
          type: boolean
          enum:
            - false
      required:
        - message
        - statusCode
        - isRetryable
      description: Authentication error (401)
      example:
        message: Authentication required
        statusCode: 401
        isRetryable: false
    AuthorizationError:
      type: object
      properties:
        message:
          type: string
        statusCode:
          type: number
          enum:
            - 403
        isRetryable:
          type: boolean
          enum:
            - false
      required:
        - message
        - statusCode
        - isRetryable
      description: Authorization error (403)
      example:
        message: Access denied
        statusCode: 403
        isRetryable: false
    NotFoundError:
      type: object
      properties:
        message:
          type: string
        statusCode:
          type: number
          enum:
            - 404
        isRetryable:
          type: boolean
          enum:
            - false
      required:
        - message
        - statusCode
        - isRetryable
      description: Resource not found error (404)
      example:
        message: Resource not found
        statusCode: 404
        isRetryable: false
...
    InternalServerError:
      type: object
      properties:
        message:
          type: string
        statusCode:
          type: number
          enum:
            - 500
        isRetryable:
          type: boolean
          enum:
            - true
      required:
        - message
        - statusCode
        - isRetryable
      description: Internal server error (500)
      example:
        message: Internal server error
        statusCode: 500
        isRetryable: true
...
paths:
...
  /api/accounts/{accountId}/seasons/{seasonId}/teams/{teamSeasonId}/managers:
    post:
      description: Add a manager to the team for the given season
      operationId: addManager
      security:
        - bearerAuth: []
      tags:
        - Managers
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: number
        - name: seasonId
          in: path
          required: true
          schema:
            type: string
            format: number
        - name: teamSeasonId
          in: path
          required: true
          schema:
            type: string
            format: number
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamManager"
      responses:
        "200":
          description: Manager added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamManager"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationError"
        "403":
          description: Access denied - Account admin required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationError"
        "404":
          description: Manager not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
  /api/accounts/{accountId}/seasons/{seasonId}/teams/{teamSeasonId}/managers/{managerId}:
    delete:
      description: Remove a manager from the team for the given season
      operationId: removeManager
      security:
        - bearerAuth: []
      tags:
        - Managers
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: number
        - name: seasonId
          in: path
          required: true
          schema:
            type: string
            format: number
        - name: teamSeasonId
          in: path
          required: true
          schema:
            type: string
            format: number
        - name: managerId
          in: path
          required: true
          schema:
            type: string
            format: number
      responses:
        "200":
          description: Manager removed
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationError"
        "403":
          description: Access denied - Account admin required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationError"
        "404":
          description: Manager not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"